Website modernization
The website hasn't received anything in the last *year*. I've heard from many people that they thought the game was dead because the last PR was a year ago. Not anymore!

Lilten's weekly vlogs are now displayed on the front page, as well as on a new "updates" tab. This uses Hugo's taxonomy system.

There is now a LIVE list of online servers, fetched dynamically using JS. This made me pull in petite-vue and figure out how the hell to set that up. Anyways, the site is being exported to JS modules from typescript now. No complex JS build systems here.

I backfilled 3 of lilten's videos just so the front page doesn't seem too empty. More in the future!